requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.3.0
    # via jsonschema
aws-xray-sdk==2.6.0
    # via okdata-maskinporten-api (setup.py)
boto3==1.18.38
    # via okdata-maskinporten-api (setup.py)
botocore==1.21.38
    # via
    #   aws-xray-sdk
    #   boto3
    #   s3transfer
certifi==2020.12.5
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.4
    # via requests
cryptography==3.4.8
    # via pyopenssl
ecdsa==0.14.1
    # via python-jose
fastapi==0.68.0
    # via okdata-maskinporten-api (setup.py)
future==0.18.2
    # via aws-xray-sdk
idna==2.10
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonpickle==1.5.0
    # via aws-xray-sdk
jsonschema==3.2.0
    # via okdata-sdk
mangum==0.12.2
    # via okdata-maskinporten-api (setup.py)
okdata-aws==0.3.0
    # via okdata-maskinporten-api (setup.py)
okdata-resource-auth==0.1.3
    # via okdata-maskinporten-api (setup.py)
okdata-sdk==0.6.0
    # via okdata-aws
prettytable==2.0.0
    # via okdata-sdk
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pydantic==1.7.4
    # via
    #   fastapi
    #   okdata-aws
    #   okdata-maskinporten-api (setup.py)
pyjwt==2.0.1
    # via
    #   okdata-maskinporten-api (setup.py)
    #   okdata-sdk
pyopenssl==20.0.1
    # via okdata-maskinporten-api (setup.py)
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via botocore
python-jose==3.2.0
    # via python-keycloak
python-keycloak==0.24.0
    # via
    #   okdata-maskinporten-api (setup.py)
    #   okdata-sdk
requests==2.26.0
    # via
    #   okdata-maskinporten-api (setup.py)
    #   okdata-resource-auth
    #   okdata-sdk
    #   python-keycloak
rsa==4.7
    # via python-jose
s3transfer==0.5.0
    # via boto3
simplejson==3.17.5
    # via okdata-maskinporten-api (setup.py)
six==1.15.0
    # via
    #   ecdsa
    #   jsonschema
    #   pyopenssl
    #   python-dateutil
    #   python-jose
starlette==0.14.2
    # via fastapi
structlog==20.2.0
    # via okdata-aws
typing-extensions==3.10.0.0
    # via mangum
urllib3==1.26.5
    # via
    #   botocore
    #   okdata-sdk
    #   requests
wcwidth==0.2.5
    # via prettytable
wrapt==1.12.1
    # via aws-xray-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
